Move wiki models tests to addons.wiki.tests.test_models
